refactor: Simplify AzNavRail library and fix build
- Consolidate multiple data models into a single `AzNavItem` model.
- Remove deprecated and redundant UI components.
- Refactor the main `AzNavRail` to use an internal `AzNavRailUi`.
- Fix incorrect use of `remember(text)` in `AzNavRailButton`.
- Update Kotlin and AGP versions.
- Add `publishing` block to `build.gradle.kts` to fix artifact generation.
