+++ gradle/libs.versions.toml
@@ -6,8 +6,8 @@ material3 = "1.4.0-beta01"
 coil = "2.6.0"
 
 # Plugins
-kotlin = "2.0.0"
-androidLibrary = "8.6.0"
+kotlin = "2.2.10"
+androidLibrary = "8.12.2"
 activityCompose = "1.9.0"
 junit = "4.13.2"
 androidx-test-ext-junit = "1.1.5"
